+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Validation for `workers` count. Worker count should be a positive integer.
+
+### Changed
+
 - Rename module `Rapidflow` to `RapidFlow`.
 
 ## [0.1.0] - 2025.11.01